Moved mplgraphicsview to utilities folder
